td-seller-intelligence/scripts/courts/divorce_monitor.py
"""
Divorce Court Monitor for TD Realty Seller Intelligence

Monitors Franklin County Domestic Relations Court for divorce/dissolution filings.
Divorcing couples often need to sell the marital home as part of the settlement.

Data Sources:
- Franklin County Clerk of Courts: https://www.franklincountyohio.gov/clerk/
- Ohio Courts Network: https://www.courtclerk.org/records-search/
"""

import logging
import re
from datetime import date, datetime
from typing import List, Dict, Any, Optional
import requests
from bs4 import BeautifulSoup

from scripts.courts.base_monitor import BaseCourtMonitor, LifeEvent, extract_address_from_text

logger = logging.getLogger(__name__)


class DivorceMonitor(BaseCourtMonitor):
    """
    Monitor Franklin County Domestic Relations Court for divorce filings.

    Tracks:
    - Divorce filings
    - Dissolution of marriage
    - Legal separations (may precede divorce)

    These cases often involve property division where the marital home is sold.
    """

    EVENT_TYPE = 'DIVORCE'
    COURT_NAME = 'Franklin County Domestic Relations'

    # Franklin County Clerk search endpoints
    BASE_URL = 'https://www.franklincountyohio.gov'
    SEARCH_URL = 'https://drpublic.fccourts.org'

    # Case type codes for domestic relations
    CASE_TYPES = ['DR', 'DM', 'DS']  # Divorce, Dissolution, Legal Separation

    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        })

    def fetch_new_cases(self, start_date: date, end_date: date) -> List[LifeEvent]:
        """
        Fetch new divorce/dissolution filings from Franklin County.

        Args:
            start_date: Start of date range.
            end_date: End of date range.

        Returns:
            List of LifeEvent objects for new filings.
        """
        events = []

        try:
            # Try the public records API first
            api_events = self._fetch_from_api(start_date, end_date)
            events.extend(api_events)
        except Exception as e:
            logger.warning(f"API fetch failed, trying web scrape: {e}")
            try:
                # Fall back to web scraping
                scrape_events = self._fetch_from_web(start_date, end_date)
                events.extend(scrape_events)
            except Exception as e2:
                logger.error(f"Web scrape also failed: {e2}")

        return events

    def _fetch_from_api(self, start_date: date, end_date: date) -> List[LifeEvent]:
        """
        Fetch cases from court API/public records system.

        Franklin County provides a public case search that we can query.
        """
        events = []

        # Franklin County Domestic Relations public portal
        # Note: This is a simulation - actual API endpoints would need to be discovered
        search_url = f"{self.SEARCH_URL}/CaseSearch"

        for case_type in self.CASE_TYPES:
            try:
                params = {
                    'caseType': case_type,
                    'filedDateFrom': start_date.strftime('%m/%d/%Y'),
                    'filedDateTo': end_date.strftime('%m/%d/%Y'),
                    'pageSize': 100
                }

                logger.info(f"Searching for {case_type} cases from {start_date} to {end_date}")

                response = self.session.get(search_url, params=params, timeout=30)

                if response.status_code == 200:
                    cases = self._parse_api_response(response.text, case_type)
                    events.extend(cases)
                    logger.info(f"Found {len(cases)} {case_type} cases")
                else:
                    logger.warning(f"API returned status {response.status_code}")

            except requests.RequestException as e:
                logger.warning(f"Request failed for {case_type}: {e}")

        return events

    def _parse_api_response(self, html: str, case_type: str) -> List[LifeEvent]:
        """Parse HTML response from case search."""
        events = []
        soup = BeautifulSoup(html, 'lxml')

        # Look for case result rows
        # Structure varies by court system - this handles common patterns
        case_rows = soup.select('tr.case-row, tr.search-result, .case-item')

        for row in case_rows:
            try:
                event = self._parse_case_row(row, case_type)
                if event:
                    events.append(event)
            except Exception as e:
                logger.debug(f"Failed to parse row: {e}")

        return events

    def _parse_case_row(self, row, case_type: str) -> Optional[LifeEvent]:
        """Parse a single case row from search results."""
        # Extract case number
        case_num_elem = row.select_one('.case-number, td:first-child a, .caseNumber')
        if not case_num_elem:
            return None

        case_number = case_num_elem.get_text(strip=True)

        # Extract parties
        parties_elem = row.select_one('.parties, .case-parties, td:nth-child(2)')
        parties = parties_elem.get_text(strip=True) if parties_elem else ''

        # Extract filing date
        date_elem = row.select_one('.file-date, .filedDate, td:nth-child(3)')
        date_str = date_elem.get_text(strip=True) if date_elem else ''

        # Parse date
        event_date = self._parse_date(date_str)
        if not event_date:
            event_date = date.today()

        # Try to extract address if present
        address = None
        full_text = row.get_text()
        address = extract_address_from_text(full_text)

        # Build case URL
        case_url = None
        link = case_num_elem.get('href') if case_num_elem.name == 'a' else None
        if link:
            case_url = f"{self.SEARCH_URL}{link}" if link.startswith('/') else link

        return LifeEvent(
            event_type=self.EVENT_TYPE,
            event_date=event_date,
            case_number=case_number,
            court=self.COURT_NAME,
            parties=parties,
            address=address,
            source_url=case_url,
            raw_data={
                'case_type': case_type,
                'html_text': full_text[:500]  # First 500 chars for debugging
            }
        )

    def _fetch_from_web(self, start_date: date, end_date: date) -> List[LifeEvent]:
        """
        Fallback web scraping method.

        Scrapes the court's public search page directly.
        """
        events = []

        # Alternative: use Ohio Courts Network
        ocn_url = 'https://www.courtclerk.org/records-search/'

        try:
            logger.info("Attempting Ohio Courts Network search...")

            # Note: OCN may require CAPTCHA or login for programmatic access
            # This is a demonstration of the scraping approach

            response = self.session.get(ocn_url, timeout=30)
            if response.status_code != 200:
                logger.warning(f"OCN returned status {response.status_code}")
                return events

            soup = BeautifulSoup(response.text, 'lxml')

            # Look for a search form and submit
            form = soup.find('form', {'id': 'case-search'})
            if form:
                logger.info("Found case search form")
                # Would submit form with appropriate parameters
                pass

        except Exception as e:
            logger.error(f"Web scraping failed: {e}")

        return events

    def _parse_date(self, date_str: str) -> Optional[date]:
        """Parse date string from court records."""
        if not date_str:
            return None

        date_str = date_str.strip()

        formats = [
            '%m/%d/%Y',
            '%m-%d-%Y',
            '%Y-%m-%d',
            '%m/%d/%y',
        ]

        for fmt in formats:
            try:
                return datetime.strptime(date_str, fmt).date()
            except ValueError:
                continue

        return None


def run_divorce_monitor(
    lookback_days: int = 1,
    dry_run: bool = False
) -> Dict[str, Any]:
    """
    Run the divorce court monitor.

    Args:
        lookback_days: Days to look back for new filings.
        dry_run: If True, don't write to Sheets.

    Returns:
        Summary of the monitoring run.
    """
    monitor = DivorceMonitor(lookback_days=lookback_days, dry_run=dry_run)
    return monitor.run()


if __name__ == '__main__':
    import argparse
    import sys

    # Add parent directory to path
    sys.path.insert(0, str(__file__).rsplit('/', 3)[0])

    # Load environment variables
    from dotenv import load_dotenv
    load_dotenv()

    parser = argparse.ArgumentParser(description='Monitor Franklin County for divorce filings')
    parser.add_argument('--lookback', type=int, default=1, help='Days to look back')
    parser.add_argument('--dry-run', action='store_true', help='Do not write to Sheets')
    parser.add_argument('--verbose', '-v', action='store_true', help='Verbose output')

    args = parser.parse_args()

    if args.verbose:
        logging.basicConfig(level=logging.DEBUG)
    else:
        logging.basicConfig(level=logging.INFO)

    result = run_divorce_monitor(
        lookback_days=args.lookback,
        dry_run=args.dry_run
    )

    print(f"\nDivorce Monitor Results:")
    print(f"  Events found: {result['events_found']}")
    print(f"  Events matched: {result['events_matched']}")
    print(f"  Events saved: {result['events_saved']}")
    if result.get('errors'):
        print(f"  Errors: {result['errors']}")
